+++ buildSrc/src/main/kotlin/multiplatform-module-convention.gradle.kts
@@ -110,6 +110,13 @@ kotlin {
             jsMain.get().dependsOn(this)
             wasmJsMain.dependsOn(this)
         }
+        val nonDarwinMain by creating {
+            dependsOn(commonMain.get())
+            androidMain.get().dependsOn(this)
+            desktopMain.dependsOn(this)
+            jsMain.get().dependsOn(this)
+            wasmJsMain.dependsOn(this)
+        }
 
         val darwinMain by creating {
             dependsOn(commonMain.get())
